requirements.txt
---

## requirements.txt  

```txt
speechrecognition
keyboard
google-generativeai
pyaudio